+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Intangibility to Power Effect Builder.
 - Added Mind Link to Power Effect Builder.
 - Added Puppet to Power Effect Builder.
+- Added Slumber to Power Effect Builder.
 - Localisation for:
 - - Fear Table function.
 - - Mark Dead function.